requirement.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-tagging==0.4.6
django-taggit==1.2.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social==0.1.dev3
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
